+++ R/deps.R
@@ -3,10 +3,7 @@ addDeps <- function(tag, md, options) {
   
   # always use minified files (https://www.minifier.org)
   adminLTE_js <- "js/app.min.js"
-  custom_js <- "js/custom.js"
-  bindings_js <- "js/bindings.js"
-  sidebar_js <- "js/sidebar.js"
-  shinydashboard_js <- "shinydashboard.js"
+  shinydashboardPlus_js <- "js/shinydashboardPlus.js"
   adminLTE_css <- c("css/AdminLTE.min.css", "css/_all-skins.min.css")
   
   # material design deps
@@ -45,7 +42,7 @@ addDeps <- function(tag, md, options) {
       "shinydashboardPlus",
       as.character(utils::packageVersion("shinydashboardPlus")),
       c(file = system.file("shinydashboardPlus-0.6.0", package = "shinydashboardPlus")),
-      script = c(bindings_js, adminLTE_js, custom_js, sidebar_js),
+      script = c(adminLTE_js, shinydashboardPlus_js),
       stylesheet = adminLTE_css
     ),
     # shinydashboard css and js deps
@@ -53,7 +50,6 @@ addDeps <- function(tag, md, options) {
       "shinydashboard",
       as.character(utils::packageVersion("shinydashboard")),
       c(file = system.file(package = "shinydashboard")),
-      script = shinydashboard_js,
       stylesheet = "shinydashboard.css"
     ),
     # material design deps